IoT/v2/raspi/ML/ML code/requirements.txt
ultralytics
pyttsx3
torch
torchvision
transformers
picamera2
RPi.GPIO
